+++ changelog.md
@@ -4,10 +4,12 @@ Beta
 0.1.10 -> 0.1.11
 
 
+- Fixed titan combat
+- Fixed armored titan charging wackiness
 - Moved all UP/DOWN vector constants to ShapeUtil
+- Added "/titan stat" sub-command
 
 TODO:
-- Investigate titan combat basically not working
 - Investigate why blade handles stop swinging randomly
 - Titan endurance
 - Perfect Parrying attacks to deal counter damage (in block ability, if ticks < ~10, deal damage to attacker)